mootiro_komoo/settings/requirements.txt
Babel==0.9.6
Django==1.3.1
Fabric==1.3.4
Genshi==0.6
Markdown==2.1.1
amqplib==1.0.2
anyjson==0.3.1
argparse==1.2.1
celery==2.5.3
certifi==0.0.8
chardet==1.0.1
django-annoying==0.7.6
django-cas==2.0.3
django-celery==2.5.5
django-crispy-forms==1.1.1
django-debug-toolbar==0.9.4
django-easy-maps==0.7.2
django-markitup==1.0.0
django-nose==1.0
django-picklefield==0.2.1
django-reversion==1.5.1
django-social-auth==0.6.9
django-taggit==0.9.3
geopy==0.94.1
gunicorn==0.14.2
httplib2==0.7.4
kombu==2.1.8
lingua==1.3
mock==0.8.0
nose==1.1.2
oauth2==1.5.211
pika==0.9.5
polib==0.7.0
psycopg2==2.4.4
pycrypto==2.5
python-dateutil==1.5
python-openid==2.2.5
requests==0.11.1
simplejson==2.5.0
ssh==1.7.12
wsgiref==0.1.2
xlrd==0.7.1
xlwt==0.7.2
